// src/generated/interlink/interlink/v1/interlink_connect.ts
// @generated by protoc-gen-connect-es v1.7.0 with parameter "target=ts"
// @generated from file interlink/v1/interlink.proto (package interlink.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { BotInfo, Envelope, GetBotInfoRequest, HeartbeatRequest, HeartbeatResponse, ListBotsRequest, ListBotsResponse, RegisterBotRequest, RegisterBotResponse, SendResponse, SubscribeRequest, UnregisterBotRequest, UnregisterBotResponse } from "./interlink_pb.js";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service interlink.v1.InterlinkService
 */
export const InterlinkService = {
  typeName: "interlink.v1.InterlinkService",
  methods: {
    /**
     * @generated from rpc interlink.v1.InterlinkService.Send
     */
    send: {
      name: "Send",
      I: Envelope,
      O: SendResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc interlink.v1.InterlinkService.Subscribe
     */
    subscribe: {
      name: "Subscribe",
      I: SubscribeRequest,
      O: Envelope,
      kind: MethodKind.ServerStreaming,
    },
    /**
     * @generated from rpc interlink.v1.InterlinkService.Connect
     */
    connect: {
      name: "Connect",
      I: Envelope,
      O: Envelope,
      kind: MethodKind.BiDiStreaming,
    },
    /**
     * @generated from rpc interlink.v1.InterlinkService.RegisterBot
     */
    registerBot: {
      name: "RegisterBot",
      I: RegisterBotRequest,
      O: RegisterBotResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc interlink.v1.InterlinkService.Heartbeat
     */
    heartbeat: {
      name: "Heartbeat",
      I: HeartbeatRequest,
      O: HeartbeatResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc interlink.v1.InterlinkService.UnregisterBot
     */
    unregisterBot: {
      name: "UnregisterBot",
      I: UnregisterBotRequest,
      O: UnregisterBotResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc interlink.v1.InterlinkService.ListBots
     */
    listBots: {
      name: "ListBots",
      I: ListBotsRequest,
      O: ListBotsResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc interlink.v1.InterlinkService.GetBotInfo
     */
    getBotInfo: {
      name: "GetBotInfo",
      I: GetBotInfoRequest,
      O: BotInfo,
      kind: MethodKind.Unary,
    },
  }
} as const;